Fix DictionaryOpenHelper so that it works
Much trial and error was required to make queries to the dict table to work. The following are the changes that were applied to fixed the problems:
* incrementing DATABASE_VERSION whenever a change was made to the class
* only executing a single sql statement at a time instead of trying to run multiple statements
* dropping the dict table on upgrade
* hard-coding the table name
* changing the database name

It is unknown what actually enabled the expected behavior.

sql.txt contains the sql statments to build a sample dictionary database. It is composed 100 randomly selected rows.