Fix DictionarySearcher.katakanaToHiragana()
The long-vowel character condition to prevent its conversion to hiragana was being checked to late so that it was being converted anyway. This commit moves the condition before the check for the character being in the katakana range to prevents its erroneous conversion.

Also this removes the commented out attempt to use regular expressions to match if the given word contained any non-katakana characters. While the code is valid in Java, the JDK for Android does not appear to like unicode property blocks