Change the DictionaryEntry serialization format from JSON to ContentValue
This changes the serialization format for DictionaryEntry from JSON to ContentValues. ContentValues are a paracable object so that they can be directly placed in a bundle as an ArrayList of ContentValues. This removes the need to for the complex try/catch blocks that are required to serialze a DictionaryEntry to a JSON object and then again to serialize the list DictionaryEntries to a JSON array and the same complexity to convert back to the native objects. This change means the FloatingJapaneseDictionaryWindow does not only recieve strings as data.

Using ContentValues also has the added benefit of being able to easily convert between a CursorRow and a DictionaryEntry without the need for checking for null values.