Set the window state to open on window close
Whenever the window layout is inflated, the searchview determines whether it should be iconfied or not by the setting in the xml. As we currently have the window default to being open on start up, the search view will always be expanded whenever floating japanese dictionary is activated, regardless of the closed state. Therefore, by setting the window to open, the layout is consistent