Remove maximize button
The window maintained state information relating to maximizing independent of the layout param which meant that maximizing a collapsed window, transitioning to expanded, and unmaximizing would make the window size too small as it reverted back to the unmaximized collapsed state.

Because the UI changed, the screenshots in the readme and the readme had to be updated.