Implement search on type
This adds a onQueryTextChangeListener to the searchview and manually
calls DictionarySearcherActivity. This adds a new intent flag,
SUBMITTED, which allows DictionarySearcherActivity to differentiate
between calls that were submitted by search on type, and those
through the normal submitting process. This differentiation is
necessary because we change the query text if it is submitted
so that we can record what the voice query returned. If we did not,
we would change the query text during search on type, which breaks
IME composition.